Issue #27: Refactor injection core

+++ Tests/AutoFake.UnitTests/Setup/SourceMemberInsertMockInstallerTests.cs
@@ -21,7 +21,7 @@ namespace AutoFake.UnitTests.Setup
         [Theory, AutoMoqData]
         internal void ExpectedCalls_Func_Success(SourceMemberInsertMock mock)
         {
-            Func<byte, bool> func = x => x > 2;
+            Func<uint, bool> func = x => x > 2;
             var sut = new SourceMemberInsertMockConfiguration(mock);
 
             sut.ExpectedCalls(func);
